Add eslint-plugin-yml
The rules available to us are fairly simple. I've modified
a few rules to respect wikimedia more-spaces-are-better
policy. You can see the examples of some of the invalid
YAMLs in the tests.

One large difference with other file types is that YAML
explicitly prohibits the use of tabs as indentation, so
we standadise on 2 spaces.

Fixes #425